outputfile issue resolved
