+++ tests-system/lobster-online-report/data/expected-output.lobster
@@ -12,8 +12,8 @@
           "location": {
             "kind": "github",
             "gh_root": "https://github.com/bmw-software-engineering/lobster",
-            "commit": "e8af194598c88c4b15f3a39c125f00ea62dc7a2e",
-            "file": "test-lobster-online-report-m66px5k8/basic.py",
+            "commit": "24f75b0c9ebf1bed10d1396a441c20ce3ce15437",
+            "file": "test-lobster-online-report-idwycg78/basic.py",
             "line": 5
           },
           "name": "basic.trlc_reference",
@@ -32,8 +32,8 @@
           "location": {
             "kind": "github",
             "gh_root": "https://github.com/bmw-software-engineering/lobster",
-            "commit": "e8af194598c88c4b15f3a39c125f00ea62dc7a2e",
-            "file": "test-lobster-online-report-m66px5k8/basic.py",
+            "commit": "24f75b0c9ebf1bed10d1396a441c20ce3ce15437",
+            "file": "test-lobster-online-report-idwycg78/basic.py",
             "line": 13
           },
           "name": "basic.Example.helper_function",
@@ -50,8 +50,8 @@
           "location": {
             "kind": "github",
             "gh_root": "https://github.com/bmw-software-engineering/lobster",
-            "commit": "e8af194598c88c4b15f3a39c125f00ea62dc7a2e",
-            "file": "test-lobster-online-report-m66px5k8/basic.py",
+            "commit": "24f75b0c9ebf1bed10d1396a441c20ce3ce15437",
+            "file": "test-lobster-online-report-idwycg78/basic.py",
             "line": 17
           },
           "name": "basic.Example.nor",
